Added stub for extractUserCourses
I decided to split the quering of the subscribed users courses into
extractUserCourses and later addAllUserCourses.

I chose this approach so to make this functionality:
a) more readable
b) more flexible
This way, if a user of this library, ever needs the user information for
other purposes they can call this function, instead of having to extract
this information manually out of the self.courses.