Wrote code for addAllUserCourses
addAllUserCourses uses the ref_ids that were extracted from the ilias
urls and adds them via addCourses.

The reason why I don't simply add name and url myself to the list is to
avoid any mistakes on my side, and have only one mechanism for adding
courses in the code. This keeps the code more flexible and easier to
change.
I can also not be fully sure that the url parsed from the login_soup
won't have any random, unneeded that in it. So this keeps the
course urls consistent as well.

Code works for all courses I've tested so far, except one of cause...
Course 342184 makes problems, because it's inconsistent with other
courses formats. Problem will be addressed in next commit.