Reworked addCourse
addCourse tried to find the course name statically. Some courses use a
slightly different breadcrumb pattern, not sure how intended this is.

The course name is now found dynamically via the <a> that contains it
and points back onto the course page.

This fixes the problem mentioned in 9867ec82248dec929e2030ad3a87b481b6387717
and makes the parsing more robust.