Reworked video filename system
The ilias is inconsistent again and not every video has caption.
So we have to work with the basename in that case instead.

Taking the basename and just removing the last extension seems like the
best option for now.

This system won't be perfect. People sometimes like to use . as a word
seperator in filenames. If a file had no extension and used dots as word
seperators the last word would get swallowed.
Other problem would be if the file base name had multiple file
extensions(If the file eg got converte and they just added an
addiotional extension to the filename).
People are incosistent, so there isn't a perfect solution.
I'd call my current approach a good compromise between removing too much
or too little.